\name{splitSegMid}
\alias{splitSegMid}

\title{
Split segment
}
\description{
Split a segment into two pieces.
}
\usage{
splitSegMid(x0, y0, x1, y1, f = 0.1)
}

\arguments{
  \item{x0}{
x coord of start
}
  \item{y0}{
y coord of start
}
  \item{x1}{
x coord of finish
}
  \item{y1}{
y coord of finish
}
  \item{f}{
sigma fraction for split point
}
}
\details{
split segment
}
\value{
a vector of length two of x and y coordinate of the split point
}

\author{
Barry Rowlingson
}

\keyword{ datagen}


